made openWorkspace function to /actions. Moved function from extension.ts to openWorkspace. Made a button to treeview with incorrect funtionalty still
