+++ src/extension.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import * as init from "./init";
 
-import { resetExercise, submitExercise, testExercise } from "./actions";
+import { openWorkspace, resetExercise, submitExercise, testExercise } from "./actions/actions";
 import TMC from "./api/tmc";
 import WorkspaceManager from "./api/workspaceManager";
 import Storage from "./config/storage";
@@ -23,20 +23,7 @@ export async function activate(context: vscode.ExtensionContext) {
     const currentWorkspaceFile = vscode.workspace.workspaceFile;
     const tmcWorkspaceFile = vscode.Uri.file(resources.tmcWorkspaceFilePath);
 
-    if (currentWorkspaceFile?.toString() !== tmcWorkspaceFile.toString()) {
-        console.log("Current workspace:", currentWorkspaceFile);
-        console.log("TMC workspace:", tmcWorkspaceFile);
-        if (!currentWorkspaceFile || await askForConfirmation("Do you want to open TMC workspace and close the current one?")) {
-            await vscode.commands.executeCommand("vscode.openFolder", tmcWorkspaceFile);
-            // Restarts VSCode
-        } else {
-            const choice = "Close current and open TMC Workspace";
-            await vscode.window.showErrorMessage("Please close your current workspace before using TestMyCode.",
-            ...[choice]).then((selection) => { if (selection === choice) {
-                vscode.commands.executeCommand("vscode.openFolder", tmcWorkspaceFile);
-            }});
-        }
-    }
+    openWorkspace(currentWorkspaceFile, tmcWorkspaceFile);
 
     await vscode.commands.executeCommand("setContext", "tmcWorkspaceActive", true);
